+++ packages/json-rpc/sm-provider/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Prevent the smoldot provider from entering in an infinite loop when either smoldot or the chain crashes.
+
 ## 0.1.5 - 2024-10-06
 
 - Update dependencies